requirements.txt
openai
python-telegram-bot
pandas
python-dotenv
pillow
streamlit
fastapi
uvicorn
python-multipart
